Turbopack: make hash_xxh3_hash128 return u128 (#95232)
We already switched to 128 bit hashes for many operations, these were
still using 64 bit hashes.

- make `hash_xxh3_hash128` return `u128`, just like `hash_xxh3_hash64`
returns a `u64`
- Make `Code::source_code_hash` return a u128

+++ turbopack/crates/turbopack-browser/src/ecmascript/merged/update.rs
@@ -117,7 +117,7 @@ impl MergedModuleMap {
 
     /// Returns the hash of the module with the given id, or `None` if the
     /// module is not present in any of the versions.
-    fn get(&self, id: &ModuleId) -> Option<u64> {
+    fn get(&self, id: &ModuleId) -> Option<u128> {
         for version in &self.versions {
             if let Some(hash) = version.entries_hashes.get(id) {
                 return Some(*hash);
